+++ src/protocol/packet.rs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use crate::protocol::format;
+use super::format;
 
 state_packets!(
     handshake Handshaking {
